New feature: Custom shopping lists

+++ src/main/java/pl/pelipe/shoppinglist/user/UserService.java
@@ -7,5 +7,5 @@ public interface UserService {
 
     UserEntity findByUsername(String username);
 
-    UserEntity findById(Integer id);
+    UserEntity findById(Long id);
 }
